created new superclass for file editing
