+++ src/app/container/tool-file-editor/tool-file-editor.component.html
@@ -15,7 +15,7 @@
   -->
 
 <tabset [justified]="true">
-  <tab heading="Test Parameter Files" id="dockerfileTab">
+  <tab heading="Dockerfile" id="dockerfileTab">
     <app-code-editor-list [sourcefiles]="dockerFile" [editing]="editing" [fileType]="'dockerfile'"></app-code-editor-list>
   </tab>
   <tab heading="Descriptor Files" id="descriptorFilesTab">
