+++ src/app/shared/files.ts
@@ -23,75 +23,4 @@ export class Files {
   @Input() entrypath: string;
   @Input() publicPage: boolean;
   editing = false;
-
-  /**
-   * Toggles edit mode
-   * @return
-   */
-  toggleEdit() {
-    this.editing = !this.editing;
-  }
-
-  /**
-   * Retrieves all dockerfiles from the list of sourcefiles
-   * @param  sourceFiles Array of sourcefiles
-   * @return {Array<SourceFile>}      Array of test parameter files
-   */
-  getDockerFile(sourceFiles) {
-    return sourceFiles.filter(
-      sourcefile => sourcefile.type === 'DOCKERFILE');
-  }
-
-  /**
-   * Retrieves all descriptor files from the list of sourcefiles
-   * @param  sourceFiles Array of sourcefiles
-   * @return  {Array<SourceFile>}     Array of descriptor files
-   */
-  getDescriptorFiles(sourceFiles) {
-    return sourceFiles.filter(
-      sourcefile => sourcefile.type === 'DOCKSTORE_WDL' || sourcefile.type === 'DOCKSTORE_CWL');
-  }
-
-  /**
-   * Retrieves all test parameter files from the list of sourcefiles
-   * @param  sourceFiles Array of sourcefiles
-   * @return {Array<SourceFile>}      Array of test parameter files
-   */
-  getTestFiles(sourceFiles) {
-    return sourceFiles.filter(
-      sourcefile => sourcefile.type === 'WDL_TEST_JSON' || sourcefile.type === 'CWL_TEST_JSON');
-  }
-
-  /**
-   * Common code between tools and workflows for saving
-   * @param originalSourceFiles Sourcefiles from previous version (unedited)
-   * @param newSourceFiles Current set of sourcefiles
-   * @return A list of sourcefiles for the version to be added
-   */
-  commonSaveVersion(originalSourceFiles, newSourceFiles) {
-    const sourceFilesToDelete = [];
-
-    // Deal with file renames
-    for (const originalSourceFile of originalSourceFiles) {
-      let toDelete = true;
-      for (const newSourceFile of newSourceFiles) {
-        if (newSourceFile.path === originalSourceFile.path) {
-          toDelete = false;
-          break;
-        }
-      }
-
-      if (toDelete) {
-        const sourceFileCopy = originalSourceFile;
-        sourceFileCopy.content = null;
-        sourceFilesToDelete.push(sourceFileCopy);
-      }
-    }
-
-    if (sourceFilesToDelete.length > 0) {
-      newSourceFiles = newSourceFiles.concat(sourceFilesToDelete);
-    }
-
-    return newSourceFiles;
-  }
 }
